// SmartTooling-SPA/src/app/_core/_models/best-line/bl-layout-design-by-process-dto.ts
export interface BL_Layout_Design_ProcessDTO {
    factory_id: string;
    line_id: string;
    line_type_id: string;
    model_no: string;
    process_type_id: string;
    each_process_image_before: string;
    each_process_image_after: string;
    remarks: string;
    update_by: string;
    update_time: string;

    each_process_image_before_result: string;
    each_process_image_after_result: string;
}