// SmartTooling-SPA/src/app/_core/_models/best-line/layout-by-process-dto.ts
export interface C2BLayoutByProcessDTO {
    factory_id: string;
    line_id: string;
    line_name: string;
    line_type_id: string;
    line_type_name: string;
    model_no: string;
    model_name: string;
    process_type_id: string;
    process_type_name_en: string;
    update_by: string;
    update_time: string;
}